requirements.txt
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.1.2
colorama==0.4.4
dnspython==2.2.1
Flask==2.0.1
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pipfreeze==2.0.3
pycodestyle==2.8.0
pycparser==2.21
pymongo==4.1.1
python-dotenv==0.20.0
six==1.16.0
toml==0.10.2
Werkzeug==2.1.1
